SIGNIFICANT CHANGE
The arduino compiler started throwing segmentation error to to do with lcdprint - unresolved. Code will upload to the arduino chip.

More important the compiler noted that the const byte used to represent the node address was too long. It is 6 chars inc terminator and was set to "encodr" which is one char too long. This is now fixed in this code and in the master radio which uses the node adress to communicate.