IMPORTANT changes to data types
This code has been changed due to a problem where the initial setting of 261 degrees wasn't working. The data types for the angles and ticks of the encoder are set as float and double and could possibly do with further refinement and or standardisation. Integer arithmetic was causing the problem e.g 360/ 261 = 1