changes to azcount
the azcont variable was incremented in a now redundant routine which operated on the serial port. Comms with the stepper is now via SPI, so the if statement if(stepper.available etc is now redundant and the code can be removed.

azcount is now a volatile var and incremented in the spi interrupt. The code was tested using the monitor program and encoder comms with stepper could be seen to change.