important update
The MOSFET switch had to be changed from an N channel MOS high side  to a P channel MOS low side as the power was always on. Upshot of this is that the power ON state needs the MCU pin to be LOW, so the changes in the coe in this commit reflect that.